requirements.txt
google-generativeai
python-dotenv
pandas
jupyterlab
pydantic
# PyMuPDF for PDF processing
pymupdf
pyyaml